New services, DB models, controllers and views are up!
A lot of work has been done in the past days. There are some changes in the DB layer. The project is setup to make Elasticsearch implementable. The Index page and Identity system are up and running.  Fontawesome is also added. Yeah I will have to commit more often ;d