Fixed high severity validation issues
Every method was implemented with validation for the needed objects. Pagination is now fully implemented into the all the places where it is needed.  Also I created a static Helper class to does the pagination process and to avoid code repetition by using it. Cloudinary now gives https link to image, to make the images secure.