Started working on admin panel
I am using kinda old version of AdminLTE but it is ok I guess. For now I moved the create category logic into it. I hope I make the panel look good. Also I added a method to EfRepostiory for searching with full indices(but I am not planning on using it)